+++ Start_python_server.sh
@@ -1,0 +1,2 @@
+cd ./python_api
+uv run uvicorn main:app --reload